Remove/inline LRGB<->RGB helper functions
